ref: split nested elements logic, improved router, fixed 404 logic
