+++ package.json
@@ -19,6 +19,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-router": "^7.7.1",
+    "react-router-dom": "^7.7.1",
     "tailwindcss": "^4.1.11"
   },
   "lint-staged": {
