+++ src/hooks/usePagination.ts
@@ -1,5 +1,5 @@
 import { useMemo } from 'react';
-import { useNavigate, useParams } from 'react-router';
+import { useNavigate, useParams } from 'react-router-dom';
 
 export const usePagination = (totalPages: number) => {
   const navigate = useNavigate();
